requirements.txt
pandas
psycopg2
sqlalchemy
pandas-datareader
yfinance
apache-airflow
